requirements.txt
# requirements_cpu.txt - Dependencies optimized for CPU-only usage
transformers>=4.38.0
tokenizers>=0.15.0
accelerate>=0.27.0
llama-cpp-python>=0.2.37
sentencepiece>=0.1.99
protobuf>=4.25.1
scipy>=1.11.3
numpy>=1.24.0
tqdm>=4.66.1
huggingface_hub>=0.19.4
torch>=2.1.0
einops>=0.7.0
psutil>=5.9.6
colorama>=0.4.6
requests>=2.31.0